# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Flotas API",
        "description": "API para gestión de flotas vehiculares",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Servidor local"
        },
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
        "/drivers": {
            "get": {
                "tags": [
                    "Conductores"
                ],
                "summary": "Listar todos los conductores",
                "operationId": "b9719c3d31035ffaab6d3758673bb123",
                "parameters": [
                    {
                        "name": "available",
                        "in": "query",
                        "description": "Filtrar por disponibilidad (true/false)",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "example": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de conductores obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "license_number": {
                                                        "type": "string",
                                                        "example": "A12345678"
                                                    },
                                                    "license_expiration": {
                                                        "type": "string",
                                                        "example": "2027-01-01"
                                                    },
                                                    "is_available": {
                                                        "type": "boolean",
                                                        "example": true
                                                    },
                                                    "user": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Juan Pérez"
                                                            },
                                                            "email": {
                                                                "type": "string",
                                                                "example": "juan@email.com"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Conductores"
                ],
                "summary": "Crear un nuevo conductor",
                "operationId": "ee49d7a72113e8c63774c9eb2fa4a934",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "license_number",
                                    "license_expiration"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "license_number": {
                                        "type": "string",
                                        "example": "A12345678"
                                    },
                                    "license_expiration": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2027-01-01"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Conductor creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Conductor creado exitosamente."
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "license_number": {
                                                    "type": "string",
                                                    "example": "A12345678"
                                                },
                                                "license_expiration": {
                                                    "type": "string",
                                                    "example": "2027-01-01"
                                                },
                                                "is_available": {
                                                    "type": "boolean",
                                                    "example": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Datos inválidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The user_id field is required."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}